cursor-based pruning
